Added a middleware that performs input validation error handling logic to avoid repeating said logic in multiple routes.
